Added plots that will now also render in the gui
- Added the matplotlib plots into a new window that will now display these plots in the program, in addition to writing them to file
- These new plots are interactable, and you can zoom in/out, pan, and save the new figure if so desired
- Some functions now take in an Axes object as an input, since matplotlib's axes getter sometimes fails to produce desireable plots when used

Other fixes/changes:
- Fixed an overflow error when doing rate plots by forcing orbits values in the bhm analyser to be a 64 bit integer
- Added a mask on the lego plot such that it no longer shows histogram values with a height of zero
- Added comment headers to tkinter_gui.py to better identify gui sections

+++ tools/parser.py
@@ -79,4 +79,5 @@ def parse_text_file(file_name, start_event=0, stop_event=-1): #expects a certain
 
 
             i = i+2 #skip lines
-    return np.asarray(CH),np.asarray(AMPL),np.asarray(TDC),np.asarray(TDC2),np.asarray(BX),np.asarray(ORBIT),np.asarray(RUN_NO)
+    return np.asarray(CH),np.asarray(AMPL),np.asarray(TDC),np.asarray(TDC2),np.asarray(BX),np.asarray(ORBIT, dtype=np.int64),np.asarray(RUN_NO)
+# Note: ORBIT must be kept as a int64, otherwise rate plots may fail from integer overflow, since the default is np.int32
